Add `size_bytes` metric to track size of each torrent
I'd like to be able to see stats regarding total amount being seeded,
which is not necessarily accurately reflected just through downloaded
bytes. I may be the originator of the torrent, or rtorrent may have
shut down before persisting stats, so the downloaded amount is
occasionally incorrect for some torrents.